pyproject.toml
[project]
name = "oochy"
version = "0.1.0"
description = "Cloud-native AI agent system"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.52.0",
    "boto3>=1.35.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.115.0",
    "mangum>=0.19.0",
    "uvicorn>=0.34.0",
    "pyjwt[crypto]>=2.10.0",
]
sandbox = [
    "restrictedpython>=7.4",
    "mypy>=1.14.0",
]
infra = [
    "aws-cdk-lib>=2.180.0",
    "constructs>=10.4.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "mypy>=1.14.0",
    "ruff>=0.9.0",
    "moto[s3,sqs,iot]>=5.0.0",
]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
